requirements.txt
Flask==0.10.1
Jinja2==2.7.2
Werkzeug==0.9.4
itsdangerous==0.23
Flask-SocketIO==0.3.4
greenlet==0.4.2
python-tdbus==0.9
